+++ Hankkijogbo/Hankkijogbo/Present/MyZip/View/MyZipListHeaderView.swift
@@ -7,43 +7,26 @@
 
 import UIKit
 
-final class MyZipListHeaderView: UICollectionReusableView {
+final class MyZipListHeaderView: BaseCollectionReusableView {
         
     // MARK: - UI Properties
     
     private let headerLabel = UILabel()
     
-    // MARK: - Init
+    // MARK: - Set UI
     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        setupHierarchy()
-        setupLayout()
-        setupStyle()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-private extension MyZipListHeaderView {
-    
-    // MARK: - Private Func
-    
-    func setupHierarchy() {
+    override func setupHierarchy() {
         self.addSubview(headerLabel)
     }
     
-    func setupLayout() {
+    override func setupLayout() {
         headerLabel.snp.makeConstraints {
             $0.top.equalToSuperview().inset(17)
             $0.leading.equalToSuperview().inset(22)
         }
     }
     
-    func setupStyle() {
+    override func setupStyle() {
         headerLabel.do {
             $0.attributedText = UILabel.setupAttributedText(
                 for: PretendardStyle.caption1,
